+++ package.json
@@ -12,7 +12,8 @@
     "next": "15.2.4",
     "pericium-frontend": "file:",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "styled-jsx": "^5.1.6"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
