+++ src/app/home/index.tsx
@@ -1,0 +1,8 @@
+export default function Home(){
+    return(
+        <>
+            <h1> Home</h1>
+        
+        </>
+    )
+}
