+++ src/app/home/page.tsx
@@ -1,0 +1,8 @@
+import Home from "."
+
+export default function HomePage(){
+    return(
+            <Home/>
+
+    )
+}
