--- src/app/page.tsx
@@ -1,9 +1,0 @@
-import Login from "./login/page";
-
-export default function login() {
-  return(
-    <div>
-      <Login/>
-    </div>
-  )
-}
